Use git to create black test directories
Because 'git ls-files' is used for creating the list of python
files, then the unit tests must be able to find files checked
into git.  Added the commands to store the files into git for
the 'bats' user and to remove the test directores from git at
the completion of the tests.