Pylint errors should cause make to return error
The previous logic would always return success whether pylint
found an error or not.  Reverted to shell-like logic to filter out
any non-Python files that could be passed in by the user though
the PYTHON_FILES variable or when there are no python files and
PYTHON_FILES is set to $(PROJECT_ROOT) (shame on printf's
handling of null).