Mounts TARDIGRADE_CI_PATH to the docker container
TARDIGRADE_CI_PATH has logic that will resolve to a TARDIGRADE_CI_PROJECT
directory that is one of:

* A subdirectory of the current project. This is what `make init` sets up,
when it clones the tardigrade-ci project. i.e. `/<project>/` and
`/<project>/tardigrade-ci`.

* A parent/sibling directory of the current project. This allows a dev to
have a separate clone of tardigrade-ci in a parent/sibling directory,
for either local dev/test of tardigrade-ci, or just to reduce clone-sprawl
of tardigrade-ci. i.e. `/<project>/` and `/tardigrade-ci/`.

Previously, some weird behavior could occur with the docker/run target
in the second case. If the user had a local clone of tardigrade-ci (and
had not run `make init`), then the container did not have access to the
tardigrade-ci project, and so the 'include' for the tardigrade-ci/Makefile
would fail.