Authenticates to GitHub API with GITHUB_ACCESS_TOKEN only if present
GitHub has begun scanning for and disabling access tokens in committed,
public code (even when readonly). This broke the prior mechanism of
authenticating to the GitHub API to increase rate limits.

This patch will use GITHUB_ACCESS_TOKEN if it is present in the env.
If not present, anonymous authentication will be used (which is subject
to a lower rate limit).