Restricts container test to match tardigrade-ci containers specifically
Supports use cases where the project is using a build system (CodeBuild, travis-ci)
that is itself running in a docker container. The file `.dockerenv`
exists in those environments, so the auto-init was skipped. But these
containers do not have the `/tardigrade-ci/` project contents, so
`make` would fail to find any targets. Even calling `make init` explicitly
would fail in this scenario.

This patch ensures that auto-init will run in non-tardigrade-ci containers,
to make the targets available.