Uses latest terraform-docs, finishing with a single empty newline
The prior version of terraform-docs left the trailing newline, so
that is what all our current docs have. The latest terraform-docs
trims all newlines from the end.

This isn't the first time terraform-docs has changed how it manages
terminating newline characters. This patch takes control of the newline
behavior by stripping all newlines, `echo "$(...)"`, and adding back
a single newline at the end, `; echo`. This way we match the current
tests and the readmes in all current projects, and we won't be subject
to future changes in terraform-docs and its behavior with terminating
newlines.

This patch also disables the "modules" and "resources" outputs of
terraform-docs v0.11.x, to avoid requiring further changes to project
readmes.