Creates a non-root user in docker image and passes user context when building image
Prior to this patch, files created in the docker volume mounts were
owned by root, which causes permissions issues when accessing those
files outside the container.

This was particularly evident when running the mockstack/pytest target
(which operates inside the container), followed by running the terraform/pytest
target (which operates outside the container).

With this patch, files created in the mounted filesystem will be owned by
the user running the docker container, so there should no longer be
any permissions issues.