Uses fixuid in entrypoint to fix permissions issues with bindmounts
The user that builds the tardigrade-ci container and the user that runs
the tardigrade-ci container (or containers built from it) must have the
same UID, or permissions issues will occur.

This patch uses fixuid in the tardigrade-ci entrypoint to fix this
issue. Fixuid checks the user running the container and chowns any
file or directory owned by a specified user/group in its configuration.
This ensures the user has permissions to those files and directories.