Uses override file for mock provider config with tftest
Previously, we were directly managing the provider blocks in the terraform
tests, in order to inject the endpoints and required aliases. This had
the unfortunate effect of requiring that we *not* specify the provider
block in test config. Which also meant we would need to support directly
all provider arguments we wanted to use in the test config.

This patch removes all of that. Instead, it relies on the terraform
overrides feature, to inject a provider config into the test. The provider
config in the override file is merged with any existing provider config.
This allows the test config to maintain its own provider block for the
"nomock" use case, and take advantage of any argument supported by the
provider. For the "mockstack" use case, tftest will write an override file
to the test directory. The override file contains a provider config with
the arguments and endpoints needed for the "mockstack" use case.