Ensures root Makefile targets are available even to "guard" targets
Previously, if a user defined their own target to install an application
and used the "guard" to auto-install that program as a dependency, the
installation would fail with an error about the install target not
being available.

That occurred due to the construction of the `$(SELF)` reference, and
how it always used the tardigrade-ci Makefile instead of the user's
original Makefile, e.g. `$(lastword $(MAKEFILE_LIST))`.

By switching to `$(firstword $(MAKEFILE_LIST))` we ensure that user-defined
targets are available to the "guard". This works because tardigrade-ci
relies on an "include" in the user's Makefile that imports the tardigrade-ci
targets.